fixed 32/64 bit int mismatch.
jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@680 a2d82226-602c-0410-bba7-e51b83dd10d2

+++ powstream/powstream.c
@@ -1042,7 +1042,7 @@ NextConnection:
 			char			fname[PATH_MAX];
 			char			endname[PATH_MAX];
 			char			newpath[PATH_MAX];
-			u_int64_t		arecs,nrecs;
+			u_int32_t		arecs,nrecs;
 			off_t			hlen;
 			OWPSessionHeaderRec	hdr;
 			OWPNum64		localstop;
